pyproject.toml
[project]
name = "autotel"
dynamic = ["version"]
description = "Write One, Observe Everywhere: OpenTelemetry for Python"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = [
    "opentelemetry",
    "observability",
    "tracing",
    "monitoring",
    "telemetry",
    "distributed-tracing",
    "otel",
    "instrumentation",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
    "Typing :: Typed",
]
dependencies = [
    "opentelemetry-api>=1.38.0",
    "opentelemetry-sdk>=1.38.0",
    "opentelemetry-exporter-otlp>=1.38.0",
    "pydantic>=2.12.0",
    "typing-extensions>=4.12.0",
]

[[project.authors]]
name = "Jag Reehal"
email = "jag@jagreehal.com"

[project.urls]
Homepage = "https://github.com/jagreehal/autotel-python"
Documentation = "https://github.com/jagreehal/autotel-python#readme"
Repository = "https://github.com/jagreehal/autotel-python"
Issues = "https://github.com/jagreehal/autotel-python/issues"

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.0",
    "mypy>=1.18.0",
    "fastapi>=0.121.0",
    "httpx>=0.28.0",
    "build>=1.0.0",
    "twine>=5.0.0",
]
fastapi = [
    "opentelemetry-instrumentation-fastapi>=0.47b0",
]
django = [
    "opentelemetry-instrumentation-django>=0.47b0",
]
flask = [
    "opentelemetry-instrumentation-flask>=0.47b0",
]
events = [
    "httpx>=0.28.0",
]
logging = [
    "structlog>=25.5.0",
]
pydantic-ai = [
    "pydantic-ai>=1.19.0",
    "ollama>=0.6.0",
]
all = [
    "autotel[fastapi,django,flask,events,logging,pydantic-ai]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/autotel"]

[tool.hatch.version]
path = "src/autotel/__version__.py"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/testing/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "flask.*",
    "structlog.*",
    "django.*",
]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests", "examples"]
skips = ["B101"]  # Skip assert_used (common in tests)
